+++ Changelog.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- auto_forget option: automatically mark added card as unknown
+
 
 ## [0.4.1] - 2022-08-21
 ### Fixed